+++ contracts/babylon/src/msg/contract.rs
@@ -2,7 +2,7 @@ use crate::msg::btc_header::{BtcHeader, BtcHeaderResponse, BtcHeadersResponse};
 use crate::msg::cz_header::CzHeaderResponse;
 use crate::msg::epoch::EpochResponse;
 use crate::state::config::Config;
-use babylon_apis::Bytes;
+use babylon_apis::finality_api::Evidence;
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{Binary, StdError, StdResult};
 
@@ -92,14 +92,7 @@ pub enum ExecuteMsg {
     ///
     /// This will be forwarded over IBC to the Babylon side for propagation to other Consumers, and
     /// Babylon itself
-    Slashing {
-        /// `fp_btc_pk` is the BTC PK of the slashed finality provider
-        fp_btc_pk: Bytes,
-        /// `block_height` is the Consumer blockchain slashing height
-        block_height: u64,
-        /// `secret_key` is the secret key extracted from the slashing evidence
-        secret_key: Bytes,
-    },
+    Slashing { evidence: Evidence },
 }
 
 #[cw_serde]