+++ contracts/btc-staking/src/finality.rs
@@ -288,9 +288,7 @@ fn slash_finality_provider(
     // Raises slashing event to babylon over IBC.
     // Send to babylon-contract for forwarding
     let msg = babylon_contract::ExecuteMsg::Slashing {
-        fp_btc_pk: evidence.fp_btc_pk.clone(),
-        block_height: evidence.block_height,
-        secret_key: btc_sk.to_bytes(),
+        evidence: evidence.clone(),
     };
 
     let babylon_addr = CONFIG.load(store)?.babylon;
@@ -662,7 +660,8 @@ pub(crate) mod tests {
         assert_eq!(res.height, 0);
 
         // Add a delegation, so that the finality provider has some power
-        let del1 = crate::contract::tests::get_derived_btc_delegation(1, &[pk_hex.clone()]);
+        let mut del1 = crate::contract::tests::get_derived_btc_delegation(1, &[1]);
+        del1.fp_btc_pk_list = vec![pk_hex.clone()];
 
         let msg = ExecuteMsg::BtcStaking {
             new_fp: vec![],
@@ -795,7 +794,8 @@ pub(crate) mod tests {
         let _res = execute(deps.as_mut(), initial_env.clone(), info.clone(), msg).unwrap();
 
         // Add a delegation, so that the finality provider has some power
-        let del1 = crate::contract::tests::get_derived_btc_delegation(1, &[pk_hex.clone()]);
+        let mut del1 = crate::contract::tests::get_derived_btc_delegation(1, &[1]);
+        del1.fp_btc_pk_list = vec![pk_hex.clone()];
 
         let msg = ExecuteMsg::BtcStaking {
             new_fp: vec![],
@@ -950,7 +950,8 @@ pub(crate) mod tests {
         let _res = execute(deps.as_mut(), initial_env.clone(), info.clone(), msg).unwrap();
 
         // Add a delegation, so that the finality provider has some power
-        let del1 = crate::contract::tests::get_derived_btc_delegation(1, &[pk_hex.clone()]);
+        let mut del1 = crate::contract::tests::get_derived_btc_delegation(1, &[1]);
+        del1.fp_btc_pk_list = vec![pk_hex.clone()];
 
         let msg = ExecuteMsg::BtcStaking {
             new_fp: vec![],
@@ -1046,27 +1047,11 @@ pub(crate) mod tests {
         assert_eq!(1, res.events.len());
         // Assert the slashing event is proper
         assert_eq!(res.events[0].ty, "slashed_finality_provider".to_string());
-        // Extract the secret key to compare it
-        let pk = eots::PublicKey::from_hex(&pk_hex).unwrap();
-        let btc_sk = pk
-            .extract_secret_key(
-                &pub_rand_one,
-                &add_finality_signature.block_app_hash,
-                &finality_signature,
-                &add_finality_signature_2.block_app_hash,
-                &add_finality_signature_2.finality_sig,
-            )
-            .unwrap();
         assert_eq!(
             res.messages[0],
             SubMsg::new(WasmMsg::Execute {
                 contract_addr: babylon_addr.to_string(),
-                msg: to_json_binary(&babylon_contract::ExecuteMsg::Slashing {
-                    fp_btc_pk: btc_pk.clone(),
-                    block_height: submit_height,
-                    secret_key: btc_sk.to_bytes()
-                })
-                .unwrap(),
+                msg: to_json_binary(&babylon_contract::ExecuteMsg::Slashing { evidence }).unwrap(),
                 funds: vec![]
             })
         );