+++ datagen/utils/btccheckpoint.go
@@ -1,10 +1,8 @@
 package utils
 
 import (
-	"math/rand"
 	"os"
 	"path/filepath"
-	"time"
 
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 )
@@ -14,7 +12,6 @@ const (
 )
 
 func GenRawCheckpoint(dir string) {
-	r := rand.New(rand.NewSource(time.Now().UnixNano()))
 	randomRawCkpt := datagen.GenRandomRawCheckpoint(r)
 	randomRawCkpt.EpochNum = 12345
 	randomRawCkptBytes, err := randomRawCkpt.Marshal()