+++ packages/proto/src/gen/babylon.btcstaking.v1.rs
@@ -47,6 +47,45 @@ impl BtcSigType {
         }
     }
 }
+/// FinalityProvider defines a finality provider
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct FinalityProvider {
+    /// addr is the bech32 address identifier of the finality provider.
+    #[prost(string, tag="1")]
+    pub addr: ::prost::alloc::string::String,
+    /// description defines the description terms for the finality provider.
+    #[prost(message, optional, tag="2")]
+    pub description: ::core::option::Option<cosmos_sdk_proto::cosmos::staking::v1beta1::Description>,
+    /// commission defines the commission rate of the finality provider.
+    #[prost(string, tag="3")]
+    pub commission: ::prost::alloc::string::String,
+    /// btc_pk is the Bitcoin secp256k1 PK of this finality provider
+    /// the PK follows encoding in BIP-340 spec
+    #[prost(bytes="bytes", tag="4")]
+    pub btc_pk: ::prost::bytes::Bytes,
+    /// pop is the proof of possession of the btc_pk, where the BTC
+    /// private key signs the bech32 bbn addr of the finality provider.
+    #[prost(message, optional, tag="5")]
+    pub pop: ::core::option::Option<ProofOfPossessionBtc>,
+    /// slashed_babylon_height indicates the Babylon height when
+    /// the finality provider is slashed.
+    /// if it's 0 then the finality provider is not slashed
+    #[prost(uint64, tag="6")]
+    pub slashed_babylon_height: u64,
+    /// slashed_btc_height indicates the BTC height when
+    /// the finality provider is slashed.
+    /// if it's 0 then the finality provider is not slashed
+    #[prost(uint64, tag="7")]
+    pub slashed_btc_height: u64,
+    /// sluggish defines whether the finality provider is detected sluggish
+    #[prost(bool, tag="8")]
+    pub sluggish: bool,
+    /// consumer_id is the ID of the consumer the finality provider is operating on.
+    /// If it's missing / empty, it's assumed the finality provider is operating in the Babylon chain.
+    #[prost(string, tag="9")]
+    pub consumer_id: ::prost::alloc::string::String,
+}
 /// BTCDelegation defines a BTC delegation
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]